git commit -m "Fix: Rename tutorial-progress.entity.ts and update all references"
